Switch from 301 to 302 redirects
When upgrading from http to https, we currently have a 301
Moved Permanently redirect.  This is *correct*, but a slight
mistake is very hard to resolve as the clients will cache them
for a non-deterministic amount of time.

This changes the default to be a 302, so each request to a
http address will be upgraded.

Signed-off-by: Dave Walker <email@daviey.com>